[renamed orderbook1 to orderbook][added manual test scenarios] [references issue #3]
