--- lisk-dex-electron/src/tests/fixtures/orderbook1.js
@@ -1,21 +1,0 @@
-// Price in LSK per leasehold, Amount in leaseholds
-const asks = [
-  { price: 0.9500, amount: 2000.0000 },
-  { price: 0.9000, amount: 913.4444 },
-  { price: 0.8700, amount: 66.7126 },
-  { price: 0.8500, amount: 6.4706 },
-  { price: 0.8000, amount: 10.0000 },
-  { price: 0.7700, amount: 3.1169 },
-];
-
-// Price in LSK per leasehold, Amount in lisks
-const bids = [
-  { price: 0.6000, amount: 91.7376 },
-  { price: 0.4000, amount: 62.6993 },
-  { price: 0.3000, amount: 8.7704 },
-  { price: 0.2500, amount: 15.0000 },
-  { price: 0.2000, amount: 10.0000 },
-];
-
-export { asks, bids };
-
