--- lisk-dex-electron/src/tests/fixtures/orderbook4.js
@@ -1,13 +1,0 @@
-const asks = [
-  { price: 34, amount: 89 },
-  { price: 34, amount: 89 },
-  { price: 89, amount: 76 },
-];
-
-const bids = [
-  { price: 34, amount: 89 },
-  { price: 34, amount: 89 },
-  { price: 89, amount: 76 },
-];
-
-export { asks, bids };
